// projects/xcloud-grid/src/enums/grid-topic.enum.ts
export enum GridTopicEnum {
    _RefreshQuery = '_RefreshQuery',
    _ColumnDefinition = '_ColumnDefinition',
    _ViewDefinition = '_ViewDefinition',
    _HistoryChange = '_HistoryChange',
    _History = '_History',
    _ListData = '_ListData',
    ProcessKey = 'ProcessKey',
    LinkFieldClick = 'LinkFieldClick',
    FlowProcessEdit = 'FlowProcessEdit',
    EnableRowOperation = 'EnableRowOperation',
    EnableFilterView = 'EnableFilterView',
    EnableColumnFrozen = 'EnableColumnFrozen',
    ShowNestedDataLevel = 'ShowNestedDataLevel',
    NestedToggleField = 'NestedToggleField',
    ColumnWidthChange = 'ColumnWidthChange',
    RowSelectMode = 'RowSelectMode',
    RowSelected = 'RowSelected',
    RowOperating = 'RowOperating',
    ColumnDefinition = 'ColumnDefinition',
    ViewDefinition = 'ViewDefinition',
    ListData = 'ListData',
    History = 'History',
    OpenAdvancePanel = 'OpenAdvancePanel',
    CloseAdvancePanel = 'CloseAdvancePanel',
    FilterViewCreateOrUpdate = 'FilterViewCreateOrUpdate',
    FreezeColumn = 'FreezeColumn',
    UnFreezeColumn = 'UnFreezeColumn',
    ColumnVisialSettingChange = 'ColumnVisialSettingChange'
}